Workaround to be able to display MarkupMultiline into text widget
 Temporary hack to handle MarkupLine and MarkupMultiline on EEF property
view

The current text area widget tries to resolve conflit on value by before
disposing checking the the value of the text is equal to the current
value. However the current implementation of the
MarkupMultiline.toMarkdown() some time adds a trailing \n making the
input string and the given string not equals. At some point we will need
to create a custom widget for those data type. The implementation should
take care of this limitation and then remove this preference.